pyproject.toml
[tool.poetry]
name = "auto-awesome"
version = "0.1.0"
description = ""
authors = ["alialsaeedi19 <alialsaeedi19@gmail.com>"]
readme = "README.md"
packages = [{include = "clients"}, {include = "data_pipelines"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "*"
python-dotenv = "*"
pandas = "*"
bs4 = "*"
youtube-search-python = "*"
podsearch = "*"
streamlit = "*"
streamlit-chat = "*"


[tool.poetry.group.dev.dependencies]
black = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
